pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
    "dotenv>=0.9.9",
    "gunicorn>=25.1.0",
    "psycopg2",
    "rollbar>=1.3.0",
    "ruff>=0.15.2",
]
